Delete: Offset.rb -> absorbed into Shift class. Add Offset tests to shift_test
